pyproject.toml
[project]
name = "tradex-strategy-pc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dependency-injector>=4.47.1",
    "pandas>=2.2.3",
    "pydantic-settings>=2.9.1",
    "pyyaml>=6.0.2",
    "watchdog>=6.0.0",
    "yfinance>=0.2.36",
]

[tool.uv]
dev-dependencies = [
    "prettyprint>=0.1.5",
    "pytest>=8.3.5",
]
[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [
  ".",
  "src",
]

[tool.ruff]
fix = true
preview = true
line-length = 88
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "E701", "E302", "E303", "F401", "F403", "F405", "F821",]
